Shundori/requirements.txt
python-telegram-bot==20.7
apscheduler==3.10.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3
pandas==2.1.3

